handle PR comments: code/comments moved
